Usuario casi completo con conexion a base de datos

+++ user-service/src/main/java/com/ps/user_service/User/Service/Implementation/GenderServiceImpl.java
@@ -1,12 +1,12 @@
 package com.ps.user_service.User.Service.Implementation;
 
+import com.ps.user_service.User.Dto.GenderDto;
 import com.ps.user_service.User.Models.Gender;
 import com.ps.user_service.User.Repository.GenderRepository;
 import com.ps.user_service.User.Service.Interface.IGenderService;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class GenderServiceImpl implements IGenderService {
@@ -18,30 +18,48 @@ public class GenderServiceImpl implements IGenderService {
     }
 
     @Override
-    public List<Gender> findAll() {
-        return genderRepository.findAll();
+    public GenderDto save(GenderDto gender) {
+        Gender newGender = new Gender();
+        newGender.setDescription(gender.getDescription());
+        genderRepository.save(newGender);
+
+        GenderDto response = new GenderDto();
+        response.setDescription(newGender.getDescription());
+
+        return response;
     }
 
     @Override
-    public Optional<Gender> findById(Long id) {
-        return genderRepository.findById(id);
+    public List<GenderDto> findAll() {
+        List<Gender> genders = genderRepository.findAll();
+
+        return genders.stream().map(g -> {
+            GenderDto dto = new GenderDto();
+            dto.setDescription(g.getDescription());
+            return dto;
+        }).toList();
     }
 
     @Override
-    public Gender save(Gender gender) {
-        return genderRepository.save(gender);
-    }
-
-    @Override
-    public Gender update(Long id, Gender gender) {
+    public GenderDto update(Long id, GenderDto gender) {
         Gender existing = genderRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Género no encontrado"));
         existing.setDescription(gender.getDescription());
-        return genderRepository.save(existing);
+        
+        genderRepository.save(existing);
+
+        GenderDto response = new GenderDto();
+        response.setDescription(existing.getDescription());
+        
+        return response;
     }
 
     @Override
     public void delete(Long id) {
+        if (!genderRepository.existsById(id)) {
+            throw new RuntimeException("Género no encontrado");
+        }
         genderRepository.deleteById(id);
+        
     }
 }
